feat(/users?page): adds a new GET users endpoint, and new tests are added
This commit adds a new endpoint with the GET method that returns a paginated amount (set at 15 users) of users. It takes in the Request Parameter "page" and returns a List<UserResponseDto>. Tests are added for the POST /users endpoint, with WebMvc.

Closes GET_USERS_PAGINATED
BREAKING CHANGE: a new endpoint is added to fetch users based on pagination, and tests are added for POST /users endpoint.